% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_wheelmap_locations}
\alias{extract_wheelmap_locations}
\title{Extracts data frame with Wheelmap locations from response object.}
\usage{
extract_wheelmap_locations(metadata, content_parsed)
}
\arguments{
\item{metadata}{data frame containing meta data related to the request, see extract_wheelmap_metadata.}

\item{content_parsed}{parsed content of a response to wheelmap query}
}
\value{
data frame containing sources.
}
\description{
\code{extract_wheelmap_locations} extracts a data frame containing the Wheelmap sources that
matched the request to Wheelmap location endpoint.
}
