% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_wheelmap_locations.R
\name{get_wheelmap_locations}
\alias{get_wheelmap_locations}
\title{Get list of all locations}
\usage{
get_wheelmap_locations(api_key = api_key, q = NULL, bbox = NULL,
  wheelchair = NULL, page = NULL, per_page = NULL)
}
\arguments{
\item{api_key}{Your API key}

\item{q}{Filter nodes by a search term}

\item{bbox}{Filter nodes by a bounding box (bbox=west,south,east,north) as comma separated float numbers wich are longitude, latitude values in degrees.}

\item{wheelchair}{Filter nodes by a wheelchair status (Valid values: yes, limited, no, unknown)}

\item{page}{For pagination, what page of the results you are on. Default is 1.}

\item{per_page}{For pagination, how many results to return per page. Default is 200. Max is 500.}
}
\value{
the results from the search
}
\description{
Get list of all locations
}
\examples{
\dontrun{
get_wheelmap_locations(api_key = api_key, bbox="52.3570365, 52.6770365, 13.2288599, 13.5488599") #Daten für Berlin
# Example bbox values for different cities
# Berlin "52.3570365, 52.6770365, 13.2288599, 13.5488599"
# Dortmund "51.3542273, 51.6742273, 7.3052789, 7.6252789"
# Frankfurt "50.0155435, 50.2271408, 8.4718253, 8.8004716"
# Hamburg "53.3951118, 53.7394377, 9.7304736, 10.3251401"
# München "48.0616017, 48.2482197, 11.3607802, 11.7228777"
}
}
