% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get_wheelmap_locations_all.R
\name{get_wheelmap_locations_all}
\alias{get_wheelmap_locations_all}
\title{Get list of all locations}
\usage{
get_wheelmap_locations_all(api_key = api_key, q = NULL, bbox = NULL,
  wheelchair = NULL)
}
\arguments{
\item{api_key}{Your API key}

\item{q}{Filter nodes by a search term}

\item{bbox}{Filter nodes by a bounding box (bbox=west,south,east,north) as comma separated float numbers wich are longitude, latitude values in degrees.}

\item{wheelchair}{Filter nodes by a wheelchair status (Valid values: yes, limited, no, unknown)}
}
\value{
the results from the search
}
\description{
Get list of all locations
}
\examples{
\dontrun{
get_wheelmap_locations_all(api_key = api_key, bbox="52.357, 52.677, 13.228, 13.548")
}
}
